Updates from 5/30, 5/31, 6/1
DB connection in dataSettings.json updated to the current DB host on smarterASP. UI changes, and created the "new" them under the "themes" folder in nop.Web. This is where we edit the CSS, and then run the site under that theme.. I'm pretty sure it's called "LivingGreenStL". Also, added a plugin for dropdown menus from the NopComm dashboard, but it was showing missing assembly reference, but theres some google-able resources out there for this. Not priority at the moment. Also, from wix, we can CAREFULLY get the excel or DB files for anything, and import them via the dashboard.. hopefully this comes with images, but I'm not sure. Also, sending an email to Midge, as she was mentioning something about videos she wanted added.. ALSO--- IN THE DASHBOARD FOR EDITING "TOPICS" (different views really...).. CHECK THE TOOLS TAB TO UPDATE THE ACTUAL SOURCE CODE-- JUST GOOD OL FASHIONED HTML.