Updates
Started working on the last round of updates. Fixed the layouts to 2 columns for topics.. started working on the 'octagon' slider for the left side of the column
Some thoughts about that -- the notes are in the _ColumnTwo.cshtml inside the Flexo theme, and since these are just images that cycle, all the <a> stuff applies like normal, and we could even get kind of fancy with this insofar as anyone they want to represent ORRRRR advertisements.

Overall it was nice to get the octagon to work, but this was more of a refresher/deep dive back into the NopComm structure. But it was productive on a lot of fronts.