2 column and more
Added the 2 column layout to everything but the homepage and the sign in/register pages.
Added the "ocatagon".. needs to be made responsive, but this can probably be fixed by wrapping it in a div with whatever class/name the rest of the page is running off of (master-column-wrapper I think...) but it works.
--Note about this, tried changing all the "tiles" from px to em, but I think it was a misguided idea.

Also, made a new CSS/View for making these 'more custom' layouts.. it worked great just pasta'd from VS to the "source code" panel of the admin dashboard thing.
--Thoughts on this.. granted it'd be 'correct' to use the separate CSS file, it might be worth just putting them in the bottom of the main style.css in the Flexo theme.. Have to figure out which main page (probably _Root.cshtml) is activating all the CSS/JS files... However, the 'admin dashboard' way have everything in-line, and this would probably be the easiest way to preview what it's going to look like and just leave it be. Figure it out.